refactor: create pagination util file
