+++ plugins/saved-search/versions/validation/savedSearch.js
@@ -10,7 +10,7 @@ module.exports = function createValidation (deps) {
   const {
     utils: {
       validation: { objectIdParamsSchema },
-      list: { DEFAULT_NB_RESULTS_PER_PAGE }
+      pagination: { DEFAULT_NB_RESULTS_PER_PAGE }
     }
   } = deps
