+++ src/util/list.js
@@ -1,27 +1,8 @@
 module.exports = {
 
-  getPaginationMeta,
   parseArrayValues,
   replaceBy,
 
-  DEFAULT_NB_RESULTS_PER_PAGE: 20
-
-}
-
-function getPaginationMeta ({ nbResults, page, nbResultsPerPage }) {
-  let nbPages = Math.floor(nbResults / nbResultsPerPage)
-  if (nbResults % nbResultsPerPage !== 0) {
-    nbPages += 1
-  }
-
-  const paginationMeta = {
-    nbResults,
-    nbPages,
-    page,
-    nbResultsPerPage
-  }
-
-  return paginationMeta
 }
 
 function parseArrayValues (values) {
